requirements_matching.txt
requests
sentence-transformers
numpy
scikit-learn
python-Levenshtein
pandas
tqdm